+++ docker/scripts/install-dependencies.sh
@@ -69,13 +69,17 @@ mv /wsjtx.patch ${WSJT_DIR}
 cmakebuild ${WSJT_DIR}
 rm ${WSJT_TGZ}
 
-git clone --depth 1 -b 1.5 https://github.com/wb2osz/direwolf.git
+git clone --depth 1 -b 1.6 https://github.com/wb2osz/direwolf.git
 cd direwolf
 # hamlib is present (necessary for the wsjt-x and js8call builds) and would be used, but there's no real need.
-# by setting enable_hamlib we prevent direwolf from linking to it, and it can be stripped at the end of the script.
-make enable_hamlib=
+# this patch prevents direwolf from linking to it, and it can be stripped at the end of the script.
+patch -Np1 < /direwolf-hamlib.patch
+mkdir build
+cd build
+cmake ..
+make
 make install
-cd ..
+cd ../..
 rm -rf direwolf
 # strip lots of generic documentation that will never be read inside a docker container
 rm /usr/local/share/doc/direwolf/*.pdf
@@ -106,8 +110,8 @@ popd
 rm -rf dream
 rm dream-2.1.1-svn808.tar.gz
 
-git clone https://github.com/hessu/aprs-symbols /opt/aprs-symbols
-pushd /opt/aprs-symbols
+git clone https://github.com/hessu/aprs-symbols /usr/share/aprs-symbols
+pushd /usr/share/aprs-symbols
 git checkout 5c2abe2658ee4d2563f3c73b90c6f59124839802
 popd