+++ htdocs/map.js
@@ -223,25 +223,13 @@
                     case "config":
                         var config = json.value;
                         if (!map) $.getScript("https://maps.googleapis.com/maps/api/js?key=" + config.google_maps_api_key).done(function(){
-                            var mapTypeId = config.google_maps_api_key ? 'roadmap' : 'OSM';
-
                             map = new google.maps.Map($('.openwebrx-map')[0], {
                                 center: {
                                     lat: config.receiver_gps.lat,
                                     lng: config.receiver_gps.lon
                                 },
                                 zoom: 5,
-                                mapTypeId: mapTypeId
                             });
-
-                            map.mapTypes.set("OSM", new google.maps.ImageMapType({
-                                getTileUrl: function(coord, zoom) {
-                                    return "https://maps.wikimedia.org/osm-intl/" + zoom + "/" + coord.x + "/" + coord.y + ".png";
-                                },
-                                tileSize: new google.maps.Size(256, 256),
-                                name: "OpenStreetMap",
-                                maxZoom: 18
-                            }));
 
                             $.getScript("static/lib/nite-overlay.js").done(function(){
                                 nite.init(map);