+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 **unreleased**
 - Introduced `squelch_auto_margin` config option that allows configuring the auto squelch level
+- Removed `port` configuration option; `rtltcp_compat` takes the port number with the new connectors
+- Added support for new WSJT-X modes FST4 and FST4W (only available with WSJT-X 2.3)
+- New devices supported:
+  - HPSDR devices (Hermes Lite 2)
+  - BBRF103 / RX666 / RX888 devices supported by libsddc
+  - Devices using the EB200 protocol
+
+**0.20.1**
+- Remove broken OSM map fallback
 
 **0.20.0**
 - Added the ability to sign multiple keys in a single request, thus enabling multiple users to claim a single receiver